+++ ui/src/components/tabs-generator/QasTabsGenerator.stories.js
@@ -14,8 +14,28 @@ export default {
 
   argTypes: {
     // Props
+    activeColor: {
+      descritpion: 'The color to be attributed to the text of the active tab.',
+      default: 'primary'
+    },
+
+    bgColorClass: {
+      description: 'The color class to be attributed to the background of tab.',
+      default: 'transparent'
+    },
+
+    colorClass: {
+      description: 'The color class to be attributed to the text of all tabs.',
+      default: 'text-primary'
+    },
+
     counters: {
       description: 'Number that will appear on the badge for each tab.'
+    },
+
+    indicatorColor: {
+      description: 'The color to be attributed to the indicator (the underline) of the active tab.',
+      default: 'primary'
     },
 
     tabs: {
@@ -61,7 +81,7 @@ const Template = (args, { argTypes }) => ({
   components: { QasTabsGenerator },
   props: Object.keys(argTypes),
   template:
-    '<qas-tabs-generator v-model="value" v-bind="$props" class="bg-primary text-white"/>'
+    '<qas-tabs-generator v-model="value" v-bind="$props" />'
 })
 
 export const Default = Template.bind({})